Reformat simple
Se movieron las request hacia su propio modulo de recursos, y se cambio pa que las weas esten en una ruta mas sensible, en vez de andar pajareando el el root. Aproveche y puse un log que anuncia host y puerto donde la app esta escuchando.
